etf_db/requirements.txt
alpaca-trade-api==0.52.0
asyncpg==0.21.0
psycopg2-binary==2.8.6